+++ examples/video/asynchronous_event/authenticate.cpp
@@ -60,8 +60,8 @@ int main(int argc, const char** argv) {
         .action("store_true")
         .help("GROUP A flag determining whether the enrollment ID is for an enrollment group.");
     parser.add_argument({ "-D", "--device" })
-        .default_value(0)
-        .help("DEVICE The ID of the OpenCV device to use.");
+        .default_value("0")
+        .help("DEVICE The ID of the OpenCV device to use or a path to an image / video file.");
     parser.add_argument({ "-v", "--verbose" })
         .action("store_true")
         .help("VERBOSE Produce verbose output.");
@@ -81,7 +81,7 @@ int main(int argc, const char** argv) {
     else if (args.get<std::string>("threshold") == "HIGHEST")
         THRESHOLD = RecognitionThreshold::HIGHEST;
     const auto GROUP = args.get<bool>("group");
-    const auto DEVICE = args.get<int>("device");
+    const auto DEVICE = args.get<std::string>("device");
     const auto VERBOSE = args.get<bool>("verbose");
 
     // Create an insecure credential store for keeping OAuth credentials in.
@@ -152,8 +152,9 @@ int main(int argc, const char** argv) {
 
     // Create an image capture object
     cv::VideoCapture capture;
-    if (!capture.open(DEVICE)) {
-        std::cout << "Capture from camera #" << DEVICE << " failed" << std::endl;
+    const bool IS_DEVICE_NUMERIC = !DEVICE.empty() && DEVICE.find_first_not_of("0123456789") == std::string::npos;
+    if (!(IS_DEVICE_NUMERIC ? capture.open(stoi(DEVICE)) : capture.open(DEVICE))) {
+        std::cout << "Capture from device " << DEVICE << " failed" << std::endl;
         return 1;
     }
 
@@ -194,6 +195,7 @@ int main(int argc, const char** argv) {
     std::thread event_thread([&](){
         void* tag(nullptr);
         bool ok(false);
+        bool is_running(true);
         while (queue.Next(&tag, &ok)) {
             if (!ok) continue;
             if (tag == stream) {
@@ -224,6 +226,11 @@ int main(int argc, const char** argv) {
                 std::vector<unsigned char> buffer;
                 {  // Lock the mutex and encode the frame with JPEG into a buffer.
                     std::lock_guard<std::mutex> lock(frame_mutex);
+                    // If the frame is empty, something went wrong, shutdown the loop.
+                    if (frame.empty()) {
+                        is_running = false;
+                        continue;
+                    }
                     cv::imencode(".jpg", frame, buffer);
                 }
                 // Create the request from the encoded image data.
@@ -242,6 +249,8 @@ int main(int argc, const char** argv) {
                     is_authenticated = is_authenticated && stream->getResponse().isalive();
                 score = stream->getResponse().score();
                 is_live = stream->getResponse().isalive();
+                // If the write thread has encountered an error, stop reading and exit.
+                if (!is_running) break;
                 // If we're finished enrolling, don't issue a new read request.
                 if (!is_authenticated)
                     stream->getCall()->Read(&stream->getResponse(), (void*) Events::Read);